pyproject.toml
[project]
name = "importar_csv_odoo"
version = "0.1.0"
description = "Pipeline unificado de importación CSV a Odoo v14 con adaptadores por proveedor"
requires-python = ">=3.10"
dependencies = [
  "pdfplumber>=0.11,<1",
  "pandas>=2.1,<3",
  "python-dotenv>=1",
  "PyYAML>=6",
  "loguru>=0.7",
]
